Move RemoteObject helpers to standalone file
Resolves https://github.com/grafana/xk6-browser/pull/130#discussion_r757004597
